On the way to a InstRO 4 Clang.
Added more basic files for the overall InstRO 5 design.
Changed access functions for the Clang constructs inside a ClangContructSet.
Added CygProfileAdapter files - not tested yet.
Started to build the infrastructure to actually construct an Instrumentor based
on Clang.
For that added an InstRO class and Factories. Also a main, where all comes together.
Did not integrate that into the Makefile yet.
Next steps are making the full thing compile and integrate the Autotools.