Clang Instrumentor partly functional.
This commit is a, for the first time, partly functional Clang
Instrumentor.

I had to change some things again, due to how Clang works and which
compile flags it uses. For example it does not allow RTTI, which makes
the use of dynamic_cast impossible.

Since in Clang one cannot set the ASTContext for the InstRO Passes at
construction time I worked around that using a PassExecuter object,
which first injects the AST information.
Inside the ClangPassExecuter there is a really ugly piece of code using
reinterpret_cast because for now I simply wanted it to work.