Autotools update
Small cleanup of configure.ac & lib/Makefile.am. Fixed the handling of different flags according to conventions (e.g. include dirs to CPPFLAGS, --std=c++11 to CXXFLAGS).
Added check for rapidjson (--with-rapidjson=...) to the configure script and modified the Makefile accordingly by adding the include directory and the ConfigInstrumentor target.