Updated Autotools
Export CONFIG_USE_ROSE and CONFIG_USE_CLANG (for instro/config.h). Rewrote the check for ROSE and got rid of --enable-rose (use --with-rose without any custom prefix to achieve the same effect) and ROSE_INCLUDEDIR_NEW. Removed several AC_SUBST which are not used (variables not set and not used in Makefile).
Added -Wall to the C and CXX flags when debugging is enabled.