Fixed an error and updated to use logger...
I fixed an error that Simon pointed me to. I forgot to initialize the
decidingSelector field in the adapter classes for the Clang
implementation. I discovered another error which led to a seg fault, as
I was not setting the ASTContext correctly in the pass executer.

I also moved the output to use the logger.