Large test refactoring...
I renamed the test instrumentor to reflect what it does.
Refactored the test infrastructure to support for listing the target
files a specific testinstrumentor is applied to explicitly. Each
testinstrumentor now has its own config directory under
test/input/theInstrumentor which contains a targets.lst as well as the
already known *.in files.