Updated tests....
Added another test for control structures, which failes due to ROSE's
inability to generate correct line/col information for variable
declarations within an if-condition.
Clarified expectations for template functions/memberfunction. Test fails
for member function as ROSE does not (yet?) generate an AST for member
functions but only their instantiations. Thus, InstRO cannot find the
expected constructs as they do not exist within the AST. But I strongly
argue we should expect them!
An AssignInitializer is now an expected expression, as it results in
some assembly instruction. However, a ctor-initializer is not an
observable expression and thus not a construct.