Added callpath test case for the FunctionWrapperTest
Applies both definition and wrapper prefixes as well as the renaming selector (although using the renaming selector here is not required). If test codes were allowed to fail linking, it would be better to exclude function 'callsLeafOne' from renaming in order to check whether the renaming process is only applied to descendants of the selected renaming nodes.